bodytextcomponent + headings renamed
